+++ minitwit-frontend/src/Util/Types.tsx
@@ -2,3 +2,10 @@ export interface User {
     userName: string;
     userID: number;
 }
+
+export interface IMessage {
+    image: string;
+    userName: string;
+    text: string;
+    date: string;
+}
