+++ minitwit-frontend/src/components/Timeline/Timeline.css
@@ -1,0 +1,23 @@
+.timeline-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: #282c34;
+    padding: 20px 0 100px;
+}
+
+.timeline {
+    width: 60%;
+    min-width: 270px;
+}
+
+.no-message-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.no-message-headline {
+    font-size: 30px;
+    margin: 0;
+}
