+++ minitwit-frontend/src/pages/User/UserPage.css
@@ -1,0 +1,7 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+}
